update: ウィジェットを分割する
